GradientStopsEditorを見やすくした
