+++ src/Beutl/Views/Editors/GradientStopsEditor.axaml
@@ -6,12 +6,13 @@
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              xmlns:ui="using:FluentAvalonia.UI.Controls"
              xmlns:vm="using:Beutl.ViewModels.Editors"
+             Padding="4,0"
              d:DesignWidth="300"
              x:CompileBindings="True"
              x:DataType="vm:GradientStopsEditorViewModel"
              mc:Ignorable="d">
     <Grid ColumnDefinitions="*,Auto" RowDefinitions="Auto,Auto,Auto">
-        <TextBlock Margin="8,0,0,0"
+        <TextBlock Margin="4,0,0,0"
                    HorizontalAlignment="Left"
                    VerticalAlignment="Top"
                    Text="{Binding Header}" />
@@ -19,7 +20,7 @@
         <Border x:Name="border"
                 Grid.Row="1"
                 Height="20"
-                Margin="8,0"
+                Margin="4,0"
                 CornerRadius="10">
             <Border.Background>
                 <LinearGradientBrush GradientStops="{Binding Stops}" StartPoint="0%,50%" EndPoint="100%,50%" />
@@ -28,7 +29,7 @@
         <ItemsControl x:Name="items"
                       Grid.Row="1"
                       Height="20"
-                      Margin="8,0"
+                      Margin="4,0"
                       ClipToBounds="False"
                       IsEnabled="{Binding CanEdit.Value}"
                       ItemsSource="{Binding Stops}">
@@ -75,8 +76,8 @@
                     IsEnabled="{Binding SelectedItem.Value, Converter={x:Static ObjectConverters.IsNotNull}}"
                     Orientation="Horizontal"
                     Spacing="4">
-            <Button Width="35"
-                    Height="35"
+            <Button Width="37"
+                    Height="37"
                     Click="Delete_Click">
                 <ui:SymbolIcon FontSize="16" Symbol="Delete" />
             </Button>
